+++ tests/xoto3/dynamodb/update/versioned_test.py
@@ -1,15 +1,16 @@
 import typing as ty
 
-from botocore.exceptions import ClientError
 import pytest
+from botocore.exceptions import ClientError
 
-from xoto3.dynamodb.types import ItemKey, AttrDict, Item, TableResource
-from xoto3.dynamodb.update.versioned import (
-    versioned_diffed_update_item,
-    VersionedUpdateFailure,
-    DEFAULT_MAX_ATTEMPTS_BEFORE_FAILURE,
-)
 import xoto3.dynamodb.update.versioned as xdv
+from xoto3.dynamodb.exceptions import ItemNotFoundException, get_item_exception_type
+from xoto3.dynamodb.types import AttrDict, Item, ItemKey, TableResource
+from xoto3.dynamodb.update.versioned import (
+    DEFAULT_MAX_ATTEMPTS_BEFORE_FAILURE,
+    VersionedUpdateFailure,
+    versioned_diffed_update_item,
+)
 
 xdv.MAX_TRANSACTION_SLEEP = 0.0  # no sleeps for the test
 
@@ -206,7 +207,7 @@ def test_string_item_version_coercion_to_int():
 
 def test_update_only_versioning_expression():
     assert xdv.versioned_item_expression(2, id_that_exists="id") == dict(
-        ExpressionAttributeNames={"#itemVersion": "item_version", "#idThatExists": "id"},
+        ExpressionAttributeNames={"#itemVersion": "item_version", "#idThatExists": "id",},
         ExpressionAttributeValues={":curItemVersion": 2},
         ConditionExpression="#itemVersion = :curItemVersion OR ( attribute_not_exists(#itemVersion) AND attribute_exists(#idThatExists) )",
     )
@@ -231,3 +232,15 @@ def test_make_prefetched_get_item():
     assert prefetched_getter(None, dict()) == item
     assert prefetched_getter(None, dict()) == dict(got=True)
     assert prefetched_getter(None, dict()) == dict(got=True)
+
+
+def test_nicename_getter(integration_test_id_table):
+    TestItemNotFoundException = get_item_exception_type("TestItem", ItemNotFoundException)
+
+    def no_up(item):
+        return item
+
+    with pytest.raises(TestItemNotFoundException):
+        versioned_diffed_update_item(
+            integration_test_id_table, no_up, dict(id="should-never-exist"), nicename="TestItem",
+        )