+++ xoto3/dynamodb/put.py
@@ -6,23 +6,27 @@ items to make them acceptable to boto3/DynamoDB.
 
 If you need/wish to customize this behavior, look at .prewrite.
 """
-from typing import Tuple, Optional, Union
 from logging import getLogger
+from typing import Optional, Tuple, Union
 
 from xoto3.errors import catch_named_clienterrors
 
 from .conditions import item_not_exists
+from .constants import DEFAULT_ITEM_NAME
 from .exceptions import ItemAlreadyExistsException, get_item_exception_type
-from .types import InputItem, TableResource, Item
-from .prewrite import dynamodb_prewrite
-from .utils.table import table_primary_keys, extract_key_from_item
 from .get import strongly_consistent_get_item
+from .prewrite import dynamodb_prewrite
+from .types import InputItem, Item, TableResource
+from .utils.table import extract_key_from_item, table_primary_keys
 
 logger = getLogger(__name__)
 
 
-def PutItem(Table: TableResource, Item: InputItem, *, nicename="Item", **kwargs) -> InputItem:
+def PutItem(
+    Table: TableResource, Item: InputItem, *, nicename=DEFAULT_ITEM_NAME, **kwargs
+) -> InputItem:
     """Convenience wrapper that makes your item Dynamo-safe before writing."""
+    nicename = nicename or DEFAULT_ITEM_NAME
     logger.debug(f"Put{nicename} into table {Table.name}", extra=dict(json=dict(item=Item)))
     Table.put_item(Item=dynamodb_prewrite(Item), **kwargs)
     return Item
@@ -50,7 +54,7 @@ def put_unless_exists(Table: TableResource, item: InputItem) -> Tuple[Optional[E
 
 
 def put_but_raise_if_exists(
-    Table: TableResource, item: InputItem, *, nicename: str = "Item"
+    Table: TableResource, item: InputItem, *, nicename: str = DEFAULT_ITEM_NAME,
 ) -> InputItem:
     """Wrapper for put_item that raises ItemAlreadyExistsException if the item exists,
     or a custom-generated subclass thereof if you have provided a better "nicename".
@@ -58,6 +62,7 @@ def put_but_raise_if_exists(
     If successful, just returns the passed item.
 
     """
+    nicename = nicename or DEFAULT_ITEM_NAME
     already_exists_cerror, _response = put_unless_exists(Table, item)
     if already_exists_cerror:
         raise get_item_exception_type(nicename, ItemAlreadyExistsException)(
@@ -68,11 +73,13 @@ def put_but_raise_if_exists(
     return item
 
 
-def put_or_return_existing(table: TableResource, item: InputItem) -> Union[Item, InputItem]:
+def put_or_return_existing(
+    table: TableResource, item: InputItem, *, nicename: str = DEFAULT_ITEM_NAME,
+) -> Union[Item, InputItem]:
     try:
-        put_but_raise_if_exists(table, item)
+        put_but_raise_if_exists(table, item, nicename=nicename)
         return item
     except ItemAlreadyExistsException:
         return strongly_consistent_get_item(
-            table, {key: item[key] for key in table_primary_keys(table)}
+            table, {key: item[key] for key in table_primary_keys(table)}, nicename=nicename,
         )